requirements.txt

flask
gunicorn
psycopg2-binary
boto3
python-dotenv
ffmpeg-python
